+++ tests/test_campo.py
@@ -1,0 +1,6 @@
+import pytest
+from proyecto_tfg.campo import Campo
+
+def test_campo_sin_esmalte_rechazado():
+    with pytest.raises(TypeError):
+        Campo(None)
